feat: remove redundant file and revise codes due to removal of said file
